move filter fir job view into the job view
